Default directory to function name (#26)
* Default directory to function name

Often when deploying, we would find ourselves using the directory name
as the function name, so you would write something like

```sh
$ cape deploy isprime isprime
```

which is redundant. This PR changes the type of name from an argument to
a flag. The default value is the directory argument, ie you can override
the default with

```sh
$ cape deploy isprime --name is_prime
```

It also starts to use RunE rather than run so we can return errors that
will force cobra to display usage information

* use newdemo as the default url

this is the most stable environment right now