Introduce Renderer interface for formatting command output
Introduces a render package with a Renderer interface for formatting command
output. Provides a JSON renderer, a text-template renderer, and a print
renderer, as well as management for passing a renderer via a context.Context.

The new renderer is configured by the global `-o` flag, replacing it's old
behavior. Logging statements are no longer affected by the flag. Currently,
only the deploy command uses the flag for output.