Remove log and update error messages
Removes the "error reading attestation doc" logs. The failures that
resulted in that log being output could be from things like failing to
download a function, or checksum mismatch, among other things, so this
message was technically accurate but often quite misleading. The actual
error returned from the function is also logged, and this is often much
more helpful in understanding what went wrong.

Also updates the error message when a call to sdk.Run() fails to better
reflect what happened.