Improve logging and error message connecting to enclave
When connecting to an enclave to process a request, for example to run a
function, the API may redirect to a specific enclave pool URL. This
redirect is normal behavior, and is an implementation detail. The
previous logs which referenced the redirect explicitly were not helpful
to an end-user. This change removes references to that redirect code to
make the messages more meaningful to a typical end-user.