Renable cert chain validation when calling key
We do this by updating to the latest attest package which let's us
override the current time during verifying a certificate chain. We read
the NotBefore value from the attestation document Certificate and update
the current time to be NotBefore. If we don't do this then the
certificate chain verification will fail because the Attestation
document is only valid for 4 hours.